requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
chromadb>=0.4.0
beautifulsoup4>=4.9.3
requests>=2.25.1
tenacity>=8.0.1
tiktoken>=0.5.0
termcolor>=2.0.0
tavily-python>=0.3.0
scipy>=1.9.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.0.0
selenium>=4.0.0
webdriver_manager>=4.0.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
